Refactor gas sensors
